+++ tests/utilities/test_common.py
@@ -1,0 +1,86 @@
+""" Tests for pyvista.utilities.common."""
+
+import numpy as np
+import pytest
+
+from pyvista.utilities.common import _coerce_points_like_arg
+
+
+def test_coerce_point_like_arg():
+    # Test with Sequence
+    point = [1.0, 2.0, 3.0]
+    coerced_arg = _coerce_points_like_arg(point)
+    assert isinstance(coerced_arg, np.ndarray)
+    assert coerced_arg.shape == (1, 3)
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0]]))
+
+    # Test with 1D np.ndarray
+    point = np.array([1.0, 2.0, 3.0])
+    coerced_arg = _coerce_points_like_arg(point)
+    assert isinstance(coerced_arg, np.ndarray)
+    assert coerced_arg.shape == (1, 3)
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0]]))
+
+    # Test with 2D ndarray
+    point = np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]])
+    coerced_arg = _coerce_points_like_arg(point)
+    assert isinstance(coerced_arg, np.ndarray)
+    assert coerced_arg.shape == (2, 3)
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]))
+
+
+def test_coerce_point_like_arg_copy():
+
+    # Sequence is always copied
+    point = [1.0, 2.0, 3.0]
+    coerced_arg = _coerce_points_like_arg(point, copy=True)
+    point[0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0]]))
+    
+    point = [1.0, 2.0, 3.0]
+    coerced_arg = _coerce_points_like_arg(point, copy=False)
+    point[0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0]]))
+
+    # 1D np.ndarray can be copied or not
+    point = np.array([1.0, 2.0, 3.0])
+    coerced_arg = _coerce_points_like_arg(point, copy=True)
+    point[0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0]]))
+
+    point = np.array([1.0, 2.0, 3.0])
+    coerced_arg = _coerce_points_like_arg(point, copy=False)
+    point[0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[10.0, 2.0, 3.0]]))
+
+    # 2D np.ndarray can be copied or not
+    point = np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]])
+    coerced_arg = _coerce_points_like_arg(point, copy=True)
+    point[0, 0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]]))
+
+    point = np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]])
+    coerced_arg = _coerce_points_like_arg(point, copy=False)
+    point[0, 0] = 10.0
+    assert np.array_equal(coerced_arg, np.array([[10.0, 2.0, 3.0], [4.0, 5.0, 6.0]]))
+
+
+def test_coerce_point_like_arg_errors():
+    # wrong length sequence
+    with pytest.raises(ValueError):
+        _coerce_points_like_arg([1, 2])
+
+    # wrong type
+    with pytest.raises(TypeError):
+        # allow Sequence but not Iterable
+        _coerce_points_like_arg({1, 2, 3})
+
+    # wrong length ndarray
+    with pytest.raises(ValueError):
+        _coerce_points_like_arg(np.empty(4))
+    with pytest.raises(ValueError):
+        _coerce_points_like_arg(np.empty([2, 4]))
+
+    # wrong ndim ndarray
+    with pytest.raises(ValueError):
+        _coerce_points_like_arg(np.empty([1, 3, 3]))
